update generated code + tests

+++ card.go
@@ -314,9 +314,9 @@ type Card struct {
 	// (For tokenized numbers only.) The last four digits of the device account number.
 	DynamicLast4 string `json:"dynamic_last4"`
 	// Two-digit number representing the card's expiration month.
-	ExpMonth uint8 `json:"exp_month"`
+	ExpMonth int64 `json:"exp_month"`
 	// Four-digit number representing the card's expiration year.
-	ExpYear uint16 `json:"exp_year"`
+	ExpYear int64 `json:"exp_year"`
 	// Uniquely identifies this particular card number. You can use this attribute to check whether two customers who've signed up with you are using the same card number, for example. For payment methods that tokenize card information (Apple Pay, Google Pay), the tokenized number might be provided instead of the underlying card number.
 	//
 	// *Starting May 1, 2021, card fingerprint in India for Connect will change to allow two fingerprints for the same card --- one for India and one for the rest of the world.*
